refactor: replace logger usage with common.Log for consistent logging

+++ cmd/run.go
@@ -12,7 +12,6 @@ import (
 	"github.com/krau/SaveAny-Bot/config"
 	"github.com/krau/SaveAny-Bot/core"
 	"github.com/krau/SaveAny-Bot/dao"
-	"github.com/krau/SaveAny-Bot/logger"
 	"github.com/krau/SaveAny-Bot/storage"
 	"github.com/spf13/cobra"
 )
@@ -24,32 +23,32 @@ func Run(_ *cobra.Command, _ []string) {
 	quit := make(chan os.Signal, 1)
 	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
 	sig := <-quit
-	logger.L.Info(sig, ", exitting...")
-	defer logger.L.Info("Bye!")
+	common.Log.Info(sig, ", exitting...")
+	defer common.Log.Info("Bye!")
 	if config.Cfg.NoCleanCache {
 		return
 	}
-	if config.Cfg.Temp.BasePath != "" {
+	if config.Cfg.Temp.BasePath != "" && !config.Cfg.Stream {
 		for _, path := range []string{"/", ".", "\\", ".."} {
 			if filepath.Clean(config.Cfg.Temp.BasePath) == path {
-				logger.L.Error("Invalid cache dir: ", config.Cfg.Temp.BasePath)
+				common.Log.Error("无效的缓存文件夹: ", config.Cfg.Temp.BasePath)
 				return
 			}
 		}
 		currentDir, err := os.Getwd()
 		if err != nil {
-			logger.L.Error("Failed to get current dir: ", err)
+			common.Log.Error("获取工作目录失败: ", err)
 			return
 		}
 		cachePath := filepath.Join(currentDir, config.Cfg.Temp.BasePath)
 		cachePath, err = filepath.Abs(cachePath)
 		if err != nil {
-			logger.L.Error("Failed to get absolute path: ", err)
+			common.Log.Error("获取缓存绝对路径失败: ", err)
 			return
 		}
-		logger.L.Info("Cleaning cache dir: ", cachePath)
+		common.Log.Info("正在清理缓存文件夹: ", cachePath)
 		if err := os.RemoveAll(cachePath); err != nil {
-			logger.L.Error("Failed to clean cache dir: ", err)
+			common.Log.Error("清理缓存失败: ", err)
 		}
 	}
 }
@@ -59,8 +58,8 @@ func InitAll() {
 		fmt.Println("加载配置文件失败: ", err)
 		os.Exit(1)
 	}
-	logger.InitLogger()
-	logger.L.Info("正在启动 SaveAny-Bot...")
+	common.InitLogger()
+	common.Log.Info("正在启动 SaveAny-Bot...")
 	dao.Init()
 	storage.LoadStorages()
 	common.Init()
